Add NoItemView, animation, LaunchScreen

+++ TodoList/Views/ListView.swift
@@ -15,21 +15,27 @@ struct ListView: View {
     
     var body: some View {
         
-        List {
-            
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear(duration: 0.2)) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            } else {
+                List {
+                    
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear(duration: 0.2)) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                    
+                } // List
+                .listStyle(.plain)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
-        } // List
-        .listStyle(.plain)
+        }
         .navigationTitle("Todo List 📝")
         .toolbar {
             ToolbarItem(placement: .navigationBarLeading) {
@@ -39,6 +45,7 @@ struct ListView: View {
                 NavigationLink("Add", destination: AddView())
             }
         } // toolbar
+        .bold()
         
     }
